+++ changelog.md
@@ -10,6 +10,7 @@ Actual version: "0.0.0"
 * [changed] Tracking is now more continuous and not so jumpy
 * [changed] Every annotation change is now stored to spreadsheet
 * [added] Every stitch video part is now trimmed by presence of needle holder in the operating area
+* [added] Bar plot in report with metric
 
 ## 2023-12